device-types: update armada-385-db-ap
This updates the armada-385-db-ap device-type configuration file to append
the dtb to the zImage. It also cleans useless TFTP calls for DTB and
ramdisk which are not used in nfs boot and TFTP calls for DTB in ramdisk
boot.

Note that there is no ip argument in bootargs. If ip=dhcp, the kernel hangs
on DHCP requests for now.

Signed-off-by: Quentin Schulz <quentin.schulz@free-electrons.com>